requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.7
et_xmlfile==2.0.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.5.4
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
